chore: cleanup perf_evaluator code
